Move set_up_impurity_concentration_array into the impurity helpers
The module shared its dotted name with the only function it defined.
impurity_array_helpers already holds the helpers for the same array, so the
function joins them and the module goes away.

variables.yaml is regenerated by the consistency checker: the function is now
registered when core_radiator_conc imports the helpers, which is before
calc_core_seeded_impurity_concentration registers itself, so it moves to the
front of impurity_concentration's set_by list.
