+++ cfspopcon/formulas/impurities/__init__.py
@@ -1,19 +1,11 @@
 """Routines to calculate seeded impurity concentrations and the change in effective charge and dilution due to seeded and intrinsic impurities."""
 
-from . import (
-    core_radiator_conc,
-    edge_radiator_conc,
-    impurity_array_helpers,
-    impurity_charge_state,
-    set_up_impurity_concentration_array,
-    zeff_and_dilution_from_impurities,
-)
+from . import core_radiator_conc, edge_radiator_conc, impurity_array_helpers, impurity_charge_state, zeff_and_dilution_from_impurities
 
 __all__ = [
     "core_radiator_conc",
     "edge_radiator_conc",
     "impurity_array_helpers",
     "impurity_charge_state",
-    "set_up_impurity_concentration_array",
     "zeff_and_dilution_from_impurities",
 ]
