+++ cfspopcon/formulas/impurities/impurity_array_helpers.py
@@ -4,9 +4,25 @@ from typing import Any
 
 import xarray as xr
 
+from ...algorithm_class import Algorithm
 from ...named_options import (
     AtomicSpecies,
 )
+from ...unit_handling import Unitfull
+
+
+@Algorithm.register_algorithm(return_keys=["impurity_concentration"])
+def set_up_impurity_concentration_array(intrinsic_impurity_concentration: Unitfull) -> Unitfull:
+    """Set up the impurity concentration array, starting with the intrinsic impurities.
+
+    Args:
+        intrinsic_impurity_concentration: :term:`glossary link<intrinsic_impurity_concentration>`
+
+    Returns:
+        :term:`impurity_concentration`
+    """
+    impurity_concentration = intrinsic_impurity_concentration.copy()
+    return impurity_concentration
 
 
 def make_impurity_concentration_array(
